+++ packages/babel-preset-react/babel.preset.js
@@ -1,0 +1,13 @@
+module.exports = (api, { env, react, typescript }) => ({
+    "presets": [
+        ["@babel/preset-env", env],
+        [
+            "@babel/preset-react",
+            {
+                "runtime": "automatic",
+                ...react
+            }
+        ],
+        ["@babel/preset-typescript", typescript]
+    ]
+});